REF: move TyphosDeviceContainerTitle -> TyphosDisplayTitle
