+++ typhos/suite.py
@@ -6,17 +6,14 @@ from functools import partial
 from pyqtgraph.parametertree import ParameterTree
 from pyqtgraph.parametertree import parameterTypes as ptypes
 from qtpy import QtCore, QtWidgets
-from qtpy.QtCore import Qt, Signal, Slot
 from qtpy.QtWidgets import QWidget
 
 import pcdsutils.qt
 
-from . import display as typhos_display
 from . import utils, widgets
 from .display import TyphosDeviceDisplay
 from .tools import TyphosConsole, TyphosLogDisplay, TyphosTimePlot
 from .utils import TyphosBase, clean_name, flatten_tree, save_suite
-from .widgets import SubDisplay, TyphosSidebarItem
 
 logger = logging.getLogger(__name__)
 
@@ -25,10 +22,10 @@ class SidebarParameter(ptypes.Parameter):
     """
     Parameter to hold information for the sidebar
     """
-    itemClass = TyphosSidebarItem
-    sigOpen = Signal(object)
-    sigHide = Signal(object)
-    sigEmbed = Signal(object)
+    itemClass = widgets.TyphosSidebarItem
+    sigOpen = QtCore.Signal(object)
+    sigHide = QtCore.Signal(object)
+    sigEmbed = QtCore.Signal(object)
 
     def __init__(self, embeddable=None, **opts):
         super().__init__(**opts)
@@ -37,7 +34,7 @@ class SidebarParameter(ptypes.Parameter):
 
 class DeviceParameter(SidebarParameter):
     """Parameter to hold information Ophyd Device"""
-    itemClass = TyphosSidebarItem
+    itemClass = widgets.TyphosSidebarItem
 
     def __init__(self, device, subdevices=True, **opts):
         # Set options for parameter
@@ -193,8 +190,8 @@ class TyphosSuite(TyphosBase):
         # If we got here we can't find the subdisplay
         raise ValueError(f"Unable to find subdisplay {display}")
 
-    @Slot(str)
-    @Slot(object)
+    @QtCore.Slot(str)
+    @QtCore.Slot(object)
     def show_subdisplay(self, widget):
         """
         Open a display in the dock system
@@ -210,7 +207,7 @@ class TyphosSuite(TyphosBase):
         if not isinstance(widget, QWidget):
             widget = self.get_subdisplay(widget)
         # Setup the dock
-        dock = SubDisplay(self)
+        dock = widgets.SubDisplay(self)
         # Set sidebar properly
         self._show_sidebar(widget, dock)
         # Add the widget to the dock
@@ -222,13 +219,13 @@ class TyphosSuite(TyphosBase):
         # Add to layout
         self._content_frame.layout().addWidget(dock)
 
-    @Slot(str)
-    @Slot(object)
+    @QtCore.Slot(str)
+    @QtCore.Slot(object)
     def embed_subdisplay(self, widget):
         """Embed a display in the dock system"""
         # Grab the relevant display
         if not self.embedded_dock:
-            self.embedded_dock = SubDisplay()
+            self.embedded_dock = widgets.SubDisplay()
             self.embedded_dock.setWidget(QWidget())
             self.embedded_dock.widget().setLayout(QtWidgets.QVBoxLayout())
             self.embedded_dock.widget().layout().addStretch(1)
@@ -245,8 +242,8 @@ class TyphosSuite(TyphosBase):
         self.embedded_dock.widget().layout().insertWidget(widget_count - 1,
                                                           widget)
 
-    @Slot()
-    @Slot(object)
+    @QtCore.Slot()
+    @QtCore.Slot(object)
     def hide_subdisplay(self, widget):
         """
         Hide a visible subdisplay
@@ -285,7 +282,7 @@ class TyphosSuite(TyphosBase):
         else:
             widget.hide()
 
-    @Slot()
+    @QtCore.Slot()
     def hide_subdisplays(self):
         """
         Hide all open displays
@@ -425,7 +422,7 @@ class TyphosSuite(TyphosBase):
                 logger.debug("Creating new category %r ...", category)
                 group = ptypes.GroupParameter(name=category)
                 self._tree.addParameters(group)
-                self._tree.sortItems(0, Qt.AscendingOrder)
+                self._tree.sortItems(0, QtCore.Qt.AscendingOrder)
             logger.debug("Adding %r to category %r ...",
                          parameter.name(), group.name())
             group.addChild(parameter)
@@ -441,17 +438,3 @@ class TyphosSuite(TyphosBase):
             parameter.sigEmbed.connect(partial(self.embed_subdisplay,
                                                parameter))
         return parameter
-
-
-class TyphosDeviceContainerTitle(typhos_display.TyphosDisplayTitle,
-                                 widgets.TyphosDesignerMixin):
-    """
-    Standardized Typhos Device Display title
-    """
-    toggle_requested = QtCore.Signal()
-
-    def mousePressEvent(self, event):
-        if event.button() == Qt.LeftButton:
-            self.toggle_requested.emit()
-
-        super().mousePressEvent(event)
